requirements.txt
Django==2.1.3
zappa==0.47.1
djangorestframework==3.9.0
django-cors-headers==2.4.0
django-storages==1.7.1
google-api-core==1.7.0
google-cloud-speech==0.36.2
google-cloud-storage==1.13.2
zappa-django-utils==0.4.0
grpcio==1.18.0
django-constance==2.3.1
django-picklefield==2.0
sentry-sdk==0.7.2
Fabric3==1.14.post1
pydub==0.23.1
